requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.44
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.6
aiofiles==23.2.1